Revert "Merge pull request #1534 from pymedphys/unexpose-pseudonymise"
This reverts commit 72f5e72316b7761527757b5a53fd7eeadd5dcdcb, reversing
changes made to 9b16b9cbc293276422e7209c658160962405fba9.
