--- lib/pymedphys/_experimental/pseudonymisation/api.py
@@ -1,9 +1,0 @@
-# pylint: disable = unused-import
-
-
-from pymedphys._experimental.pseudonymisation.main import (
-    get_default_pseudonymisation_keywords,
-    is_valid_strategy_for_keywords,
-    pseudonymise,
-)
-from pymedphys._experimental.pseudonymisation.strategy import pseudonymisation_dispatch
