+++ lib/pymedphys/_streamlit/apps/__init__.py
@@ -1,1 +1,1 @@
-from . import metersetmap
+from . import metersetmap, pseudonymise
