+++ lib/pymedphys/docs/cli/ref/index.rst
@@ -20,3 +20,4 @@ This presents what to write into the command prompt to use that CLI command.
     trf
     icom
     pinnacle
+    pseudonymisation
