+++ lib/pymedphys/docs/lib/ref/experimental/index.rst
@@ -15,3 +15,4 @@ Experimental Modules
     :maxdepth: 1
 
     pinnacle
+    pseudonymisation
